feat!: use tfenv to manage Terraform versions
Instead of creating a single version of Terraform per container and release, instead releases will follow semver. Containers will be tagged with the release number. GitHub Action will allow user to specify the Terraform version to use, some of which are pre-installed and some are tested, but tfenv will install and use any version at runtime.

Signed-off-by: Daniel Hill <dan@mamu.co>